requirements.txt
flask
werkzeug
pandas
scikit-learn
joblib
requests
mysql-connector-python
faiss-cpu
numpy
tensorflow